+++ far/fileview.cpp
@@ -189,7 +189,7 @@ void FileViewer::DisplayObject()
   View.Show();
 }
 
-int FileViewer::VMProcess(int OpCode,void *vParam,__int64 iParam)
+__int64 FileViewer::VMProcess(int OpCode,void *vParam,__int64 iParam)
 {
   return View.VMProcess(OpCode,vParam,iParam);
 }
