+++ far/help.hpp
@@ -125,7 +125,7 @@ class Help:public Frame
     virtual int GetTypeAndName(string &strType, string &strName);
     virtual int GetType() { return MODALTYPE_HELP; }
 
-    virtual int VMProcess(int OpCode,void *vParam,__int64 iParam);
+    virtual __int64 VMProcess(int OpCode,void *vParam,__int64 iParam);
 
     static string &MkTopic(INT_PTR PluginNumber,const wchar_t *HelpTopic,string &strTopic);
 };
